fix: parse bearer tokens consistently across billing paths
